Refactor imports to use 'plugin' module instead of 'render' in multiple files
